images/requirements.txt
roboflow
ultralytics
opencv-python
pytesseract
pandas
Pillow
matplotlib
python-dotenv
PyYAML
